requirements/common.txt
boltons==0.6.4
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-appconf==1.0.1
django-autoslug==1.8.0
django-bmemcached==0.2.2
django-bower==5.0.4
django-compressor==1.5
django-cors-headers==1.1.0
django-filter==0.9.2
django-libsass==0.3
django-markdown==0.8.4
django-meta==0.3.1
django-model-utils==2.2
django-naomi==0.7
django-storages-redux==1.2.3
django-toolbelt==0.0.1
django-widget-tweaks==1.3
djangorestframework==3.1.3
fixmydjango==0.0.9
gnureadline==6.3.3
gunicorn==19.3.0
ipython==3.1.0
jsonfield==1.0.3
libsass==0.8.2
Markdown==2.6.2
psycopg2==2.6
Pygments==2.0.2
python-binary-memcached==0.24.3
python-decouple==2.3
requests==2.7.0
six==1.9.0
static3==0.5.1
termcolor==1.1.0
